move TTR to base representation module
