+++ src/pykeen/nn/__init__.py
@@ -53,12 +53,7 @@ from .modules import (
     UMInteraction,
     interaction_resolver,
 )
-from .node_piece import (
-    NodePieceRepresentation,
-    TensorTrainRepresentation,
-    TokenizationRepresentation,
-    tokenizer_resolver,
-)
+from .node_piece import NodePieceRepresentation, TokenizationRepresentation, tokenizer_resolver
 from .pyg import (
     FeaturizedMessagePassingRepresentation,
     SimpleMessagePassingRepresentation,
@@ -72,6 +67,7 @@ from .representation import (
     PartitionRepresentation,
     Representation,
     SubsetRepresentation,
+    TensorTrainRepresentation,
     TextRepresentation,
     TransformedRepresentation,
     WikidataTextRepresentation,
